Remove old JSONSchema target
