+++ .gitmodules
@@ -1,0 +1,3 @@
+[submodule "Tests/JSONValidationTests/JSON-Schema-Test-Suite"]
+	path = Tests/JSONValidationTests/JSON-Schema-Test-Suite
+	url = https://github.com/json-schema-org/JSON-Schema-Test-Suite.git
