+++ Sources/JSONSchema/Annotations/JSONType.swift
