+++ Sources/JSONSchema/Resources/draft2020-12/meta/applicator.json
