+++ Sources/JSONSchema/Validation/Errors/SchemaIssue.swift
