+++ Sources/JSONSchema/Validation/Errors/ValidationIssue.swift
