+++ Sources/JSONSchema/Validation/ValidationResult.swift
