+++ Sources/JSONSchemaClient/main.swift
@@ -1,1 +1,1 @@
-import JSONSchema  //import JSONSchema2
+import JSONSchema  //import JSONSchema
