--- Tests/JSONSchemaTests/Tags+JSONSchema.swift
@@ -1,6 +1,0 @@
-import Testing
-
-extension Tag {
-  @Tag static var annotations: Self
-  @Tag static var typeOptions: Self
-}
